Move OpenGL code into txt.sdl.c
