+++ prboom2/src/CMakeLists.txt
@@ -262,8 +262,6 @@ set(COMMON_SRC
     textscreen/txt_main.h
     textscreen/txt_sdl.c
     textscreen/txt_sdl.h
-    textscreen/txt_gl.c
-    textscreen/txt_gl.h
     textscreen/txt_utf8.c
     textscreen/txt_utf8.h
     textscreen/fonts/convert-font
